Handle lack of connection to CCM
Updated the application to not die if we can't connect to CCM. Now it
just displays a message saying so, and proceeds. I added debug-level
logging of stack traces so I could debug this, and then I broke the
password so I don't have to sit through 27 seconds of startup every
time I want to test something.