Added roll call report
I was also working on the "roll call by room" or whatever it was called,
but it's worthless.  They certainly aren't using it currently.  SKIP.

The report is theoretically capable of running by day or by week as well
as by month, but since I don't let them choose "day" for report date,
day and week are meaningless.  I may have to revise this depending on
client wishes.