Massive report rewrite to fix blank reports
I thought it would be easy to add blank reports to the saturdays by room
and weekdays by room reports, but instead I discovered I had done all of
the blank reports wrong.  I was using the report date as-is to generate
blank reports, but in the usual case, there will be no punch data for
that time period, because blank reports are future reports.

So I had to rewrite the model.load.process_data to run on a temporary
copy of the data, so that I could load punches from the previous time
period and use those to generate the blank reports for the current time
period.  This had the unfortunate side effect of making all the reports
asynchronous, meaning I had to pass in a callback to all of them

Anyway, it was unavoidable if I wanted them to be correct, and now they
work.  Yay.